chore: shuffle internal file exports
